Fix bad `setState` call warning
