+++ packages/@remirror/react/package.json
@@ -36,7 +36,8 @@
     "linaria": "^2.0.0-rc.3",
     "react-use": "^15.3.3",
     "resize-observer-polyfill": "^1.5.1",
-    "type-fest": "^0.16.0"
+    "type-fest": "^0.16.0",
+    "use-isomorphic-layout-effect": "^1.0.0"
   },
   "devDependencies": {
     "@remirror/core": "1.0.0-next.39",
