--- packages/@remirror/react/src/components/index.ts
@@ -1,1 +1,0 @@
-export * from './providers';
