+++ packages/@remirror/react/src/hooks/__tests__/editor-hooks.spec.tsx
@@ -7,9 +7,7 @@ import { ReactPreset } from '@remirror/preset-react';
 import { BoldExtension } from '@remirror/testing';
 import { act as renderAct, render, strictRender } from '@remirror/testing/react';
 
-import { RemirrorProvider } from '../../components';
-import { createReactManager } from '../../react-helpers';
-import { useManager, useRemirror } from '../editor-hooks';
+import { RemirrorProvider, createReactManager, useManager, useRemirror } from '../../..';
 
 jest.mock('@remirror/preset-react', () => {
   const actual = jest.requireActual('@remirror/preset-react');
