+++ packages/@remirror/react/src/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './core-hooks';
 export * from './editor-hooks';
+export * from './use-react-editor';
